pyproject.toml
[project]
name = "ohc-account-invoice"
version = "0.1.0"
description = "OHC账票生成MCP服务"
authors = [
    {name = "OHC Team", email = "team@ohc.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "minio>=7.2.0",
    "boto3>=1.28.0",
    # jinja2 removed (not used)
    # python-multipart removed (no file-upload endpoints)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
